+++ lib/app.dart
@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter_toggle_tab/flutter_toggle_tab.dart';
 import 'package:globox/models/enums.dart';
 import 'package:globox/models/package.dart';
-import 'package:globox/services/get_packages.service.dart';
-import 'package:globox/services/messages_service.dart';
-import 'package:globox/services/send_messages.service.dart';
+import 'package:globox/services/queries/get_packages.service.dart';
+import 'package:globox/services/internal/messages_service.dart';
+import 'package:globox/services/queries/send_messages.service.dart';
 import 'package:globox/ui/screens/list_screen.dart';
 import 'package:globox/ui/screens/map_screen.dart';
 import 'package:globox/ui/widgets/loader.dart';
