+++ lib/services/queries/delete_package.dart
